python-radar-server/requirements.txt
Flask
netCDF4
geojson
boto3
pygrib
arm-pyart
numpy
scipy
requests
matplotlib
cartopy
plotly